+++ src/asset/image/ImageAssetHint.ts
