+++ src/asset/script/ModuleLike.ts
