+++ src/platform/OperationPluginViewInfo.ts
