+++ src/platform/PlatformEventHandler.ts
