+++ src/platform/PlatformPointEvent.ts
@@ -1,4 +1,4 @@
-import { CommonOffset } from "./commons";
+import { CommonOffset } from "../commons";
 
 /**
  * ポイントイベントの種類。
