+++ src/surface/CompositeOperation.ts
