+++ src/surface/CompositeOperationString.ts
