+++ src/surface/SurfaceLike.ts
@@ -1,4 +1,4 @@
-import { CommonSize } from "./commons";
+import { CommonSize } from "../commons";
 import { RendererLike } from "./RendererLike";
 
 /**
